% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pricer.R
\name{hedger}
\alias{hedger}
\title{hedger}
\usage{
hedger(price = price, sigma = 0.16, mu = price)
}
\arguments{
\item{price}{Is the current price of the option}

\item{sigma}{Is the variance of the mean price, mu, in terms of a probability}

\item{mu}{Mean price of the stock over time}
}
\value{
integer
}
\description{
Computes deltas for a price hedger based on standard normal distribution
}
